requirements.txt
# Core
PyYAML>=6.0
pandas>=2.0
numpy>=1.24

# Demo UI
streamlit>=1.35

# Optional IBM watsonx.ai integration
ibm-watsonx-ai>=1.3
python-dotenv>=1.0

# Tests
pytest>=8.0
